# internal/swagger/api_v1.0.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "GTB Swagger",
      "description": "GTB server API",
      "contact": {
        "email": "developer@energie-ip.com"
      },
      "version": "1.0"
    },
    "servers": [
      {
        "url": "/{basePath}",
        "variables": {
          "basePath": {
            "default": "v1.0"
          }
        }
      }
    ],
    "paths": {
      "/user/login": {
        "post": {
          "tags": [
            "authentication"
          ],
          "summary": "createToken",
          "description": "User authentication",
          "operationId": "CreateToken",
          "parameters": [],
          "requestBody": {
            "description": "User credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Credential"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CreateTokenResponse"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": []
        }
      },
      "/user/info": {
        "get": {
          "tags": [
            "authentication"
          ],
          "summary": "user Information",
          "description": "User info",
          "operationId": "getUserInfo",
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UserInfo"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
      "/setup/switch/{mac}": {
        "get": {
          "tags": [
            "setup"
          ],
          "summary": "getSwitchSetup",
          "description": "Return Switch Setup configuration",
          "operationId": "GetSwitchSetup",
          "parameters": [
            {
              "name": "mac",
              "in": "path",
              "description": "Switch Mac address",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {},
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/SwitchConfig"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        },
        "delete": {
          "tags": [
            "setup"
          ],
          "summary": "removeSwitchSetup",
          "description": "Remove Switch Setup configuration It means the Switch will be blacklisted",
          "operationId": "RemoveSwitchSetup",
          "parameters": [
            {
              "name": "mac",
              "in": "path",
              "description": "Switch mac address",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {}
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
      "/setup/switch": {
        "post": {
          "tags": [
            "setup"
          ],
          "summary": "setSwitchSetup",
          "description": "Switch Setup configuration",
          "operationId": "SetSwitchSetup",
          "parameters": [],
          "requestBody": {
            "description": "Switch config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwitchConfig"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {}
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
      "/config/switch": {
        "post": {
          "tags": [
            "config"
          ],
          "summary": "setSwitchConfig",
          "description": "Switch runtime configuration",
          "operationId": "SetSwitchConfig",
          "parameters": [],
          "requestBody": {
            "description": "Switch config",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwitchConfig"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "sucessful operation",
              "headers": {}
            },
            "401": {
              "description": "Unauthorized Access",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            },
            "default": {
              "description": "unexpected error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Error"
                  }
                }
              }
            }
          },
          "deprecated": false,
          "security": [
            {
              "Authorization": []
            }
          ]
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "SwitchConfig": {
          "title": "SwitchConfig",
          "required": [
            "mac"
          ],
          "type": "object",
          "properties": {
            "mac": {
              "type": "string",
              "description": "Switch mac address"
            },
            "cluster": {
              "type": "integer",
              "description": "Associate Cluster number",
              "format": "int32"
            },
            "ip": {
              "type": "string",
              "description": "Switch IP address"
            },
            "isConfigured": {
              "type": "boolean",
              "description": "Reset the led configuration"
            },
            "friendlyName": {
              "type": "string",
              "description": "Sensor FriendlyName"
            },
            "dumpFrequency": {
              "type": "integer",
              "description": "Dump frequency by default: 1s",
              "format": "int32"
            }
          }
        },
...
        "Error": {
          "title": "Error",
          "required": [
            "code",
            "message"
          ],
          "type": "object",
          "properties": {
            "code": {
              "type": "integer",
              "format": "int32"
            },
            "message": {
              "type": "string"
            }
          }
        },
...
        "Credential": {
          "title": "Credential",
          "required": [
            "userKey"
          ],
          "type": "object",
          "properties": {
            "userKey": {
              "type": "string",
              "description": "user key"
            }
          }
        },
        "UserInfo": {
          "title": "UserInfo",
          "type": "object",
          "properties": {
            "userHash": {
              "type": "string",
              "description": "user key"
            },
            "priviledges": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "user access"
            },
            "accessGroups": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "user group access"
            }
          }
        },
        "CreateTokenResponse": {
          "title": "CreateTokenResponse",
          "type": "object",
          "properties": {
            "tokenType": {
              "type": "string",
              "description": "Session type",
              "enum": [
                "bearer"
              ]
            },
            "accessToken": {
              "type": "string",
              "description": "Session Token"
            },
            "expireIn": {
              "type": "integer",
              "format": "int32",
              "description": "token expiration in seconds"
            }
          }
        }
      },
      "securitySchemes": {
        "Authorization": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT",
            "description": "For accessing the API a valid JWT token must be passed in all the queries in the Authorization header. Authorization header : Bearer xxxxxx.yyyyyyy.zzzzzz"
        }
      }
    },
    "security": [
      {
        "Authorization": []
      }
    ],
...
  }